Fix task 3

+++ sort.h
@@ -30,7 +30,7 @@ void quick_sort(int *array, size_t size);
 void swap_node(listint_t *i, listint_t *j);
 void swap_node_top(listint_t *i, listint_t *j);
 void swap_node_end(listint_t *i, listint_t *j);
-size_t partition(int *array, size_t low, size_t high, size_t n);
-void quicker_sort(int *array, size_t low, size_t high, size_t n);
+int partition(int *array, int low, int high, size_t n);
+void quicker_sort(int *array, int low, int high, size_t n);
 
 #endif
